fix: packed binary is not self contained on linux
