+++ build.rs
@@ -1,11 +1,14 @@
 //! Build script for smolvm.
 //!
-//! Handles finding and linking libkrun.
+//! Handles macOS weak-link setup for libkrun.
+//!
+//! Linux loads libkrun at runtime with `dlopen` so packed stubs can start
+//! before bundled libraries are extracted.
 //!
 //! # Linking Options
 //!
 //! ## Dynamic (default)
-//! Requires libkrun installed on the system:
+//! macOS can weak-link against libkrun installed on the system:
 //! ```sh
 //! brew install libkrun  # macOS
 //! cargo build
@@ -36,45 +39,24 @@
 //! LIBKRUN_STATIC=/path/to/libkrun.a cargo build
 //! ```
 
-#[cfg(target_os = "linux")]
-use std::path::Path;
+#[cfg(target_os = "macos")]
 use std::path::PathBuf;
+#[cfg(target_os = "macos")]
 use std::process::Command;
 
-/// Check if a file is a Git LFS pointer (not the actual binary).
+/// Link libkrun for legacy macOS weak-link behavior.
 ///
-/// LFS pointer files start with "version https://git-lfs.github.com/spec/v1"
-/// and are small text files. This prevents the build from trying to link
-/// against LFS pointers when the actual files haven't been fetched.
-#[cfg(target_os = "linux")]
-fn is_lfs_pointer(path: &Path) -> bool {
-    // LFS pointers are small text files (typically < 200 bytes)
-    if let Ok(metadata) = std::fs::metadata(path) {
-        if metadata.len() > 500 {
-            return false; // Too large to be an LFS pointer
-        }
-    }
-
-    // Check if the file starts with the LFS version header
-    if let Ok(content) = std::fs::read_to_string(path) {
-        return content.starts_with("version https://git-lfs.github.com/spec/v1");
-    }
-
-    false
+/// Linux intentionally does not link libkrun at build time: packed stubs must
+/// be able to start before bundled libraries are extracted, so Linux VM launch
+/// paths use `dlopen` instead.
+#[cfg(target_os = "macos")]
+fn link_krun() {
+    println!("cargo:rustc-link-arg=-Wl,-weak-lkrun");
 }
 
-/// Check if a library is available on the system via pkg-config.
+#[cfg(target_os = "macos")]
 fn has_library(name: &str) -> bool {
     pkg_config::Config::new().probe(name).is_ok()
-}
-
-/// Link libkrun — weak on macOS so the binary can start without it
-/// (packed binary mode uses dlopen instead of link-time symbols).
-fn link_krun() {
-    #[cfg(target_os = "macos")]
-    println!("cargo:rustc-link-arg=-Wl,-weak-lkrun");
-    #[cfg(not(target_os = "macos"))]
-    println!("cargo:rustc-link-lib=krun");
 }
 
 fn main() {
@@ -96,11 +78,11 @@ fn main() {
         );
     }
 
-    #[cfg(any(target_os = "macos", target_os = "linux"))]
+    #[cfg(target_os = "macos")]
     link_libkrun();
 }
 
-#[cfg(any(target_os = "macos", target_os = "linux"))]
+#[cfg(target_os = "macos")]
 fn link_libkrun() {
     println!("cargo:rerun-if-env-changed=LIBKRUN_STATIC");
     println!("cargo:rerun-if-env-changed=LIBKRUN_BUNDLE");
@@ -288,7 +270,7 @@ fn link_libkrun() {
 /// Build libkrun from the vendored submodule.
 ///
 /// Returns the path to the directory containing the built library.
-#[cfg(any(target_os = "macos", target_os = "linux"))]
+#[cfg(target_os = "macos")]
 fn build_libkrun_from_submodule() -> Option<PathBuf> {
     let manifest_dir = PathBuf::from(std::env::var("CARGO_MANIFEST_DIR").ok()?);
     let libkrun_dir = manifest_dir.join("libkrun");
